Add files via upload
I have created the paths for the other two enemies as well. I have also added the lives system where the player loses a life on colliding with the enemy and is reset back to its starting position. I have also set up the mechanics where the player can eat ghosts and the ghost are reset back to the safe house